# docs/OpenAPI/openapi.yaml
openapi: 3.0.2
info:
  title: AI Kids Care API
  version: 2.0.4
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: http://localhost:8080/api
...
paths:
...
  /v1/users:
    get:
      tags:
        - Users
      operationId: listUsers
      summary: List users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: size
          in: query
          schema:
            type: integer
            format: int32
            minimum: 1
        - name: sort
          in: query
          schema:
            type: string
          description: e.g. created_at,desc
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageOfUsers'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    post:
      tags:
        - Users
      operationId: createUsers
      summary: Create users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsersCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '409':
          $ref: '#/components/responses/ResourceConflictResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /v1/users/{user_id}:
    get:
      tags:
        - Users
      operationId: getUsers
      summary: Get users by user_id
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        '404':
          $ref: '#/components/responses/ResourceNotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    put:
      tags:
        - Users
      operationId: updateUsers
      summary: Update users by user_id
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsersUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/ResourceNotFoundResponse'
        '409':
          $ref: '#/components/responses/ResourceConflictResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    delete:
      tags:
        - Users
      operationId: deleteUsers
      summary: Delete users by user_id
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
        '404':
          $ref: '#/components/responses/ResourceNotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        traceId:
          type: string
...
    Users:
      type: object
      properties:
        user_id:
          type: integer
          format: int64
          description: 사용자 계정 ID
        login_id:
          type: string
          description: 로그인 ID(사용자 식별자)
        email:
          type: string
          description: 사용자 이메일
        phone:
          type: string
          description: 사용자 휴대폰 번호
        password_hash:
          type: string
          description: 비밀번호 해시
        status:
          type: string
          enum:
            - ACTIVE
            - PENDING
            - DISABLED
          description: 계정 상태
        last_login_at:
          type: string
          format: date-time
          description: 마지막 로그인 일시
        created_at:
          type: string
          format: date-time
          description: 생성 일시
        updated_at:
          type: string
          format: date-time
          description: 수정 일시
      required:
        - user_id
    PageOfUsers:
      type: object
      required:
        - items
        - page
        - size
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Users'
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        total:
          type: integer
          format: int64
          minimum: 0
    UsersCreateRequest:
      type: object
      properties:
        login_id:
          type: string
          description: 로그인 ID(사용자 식별자)
        email:
          type: string
          description: 사용자 이메일
        phone:
          type: string
          description: 사용자 휴대폰 번호
        password_hash:
          type: string
          description: 비밀번호 해시
        status:
          type: string
          enum:
            - ACTIVE
            - PENDING
            - DISABLED
          description: 계정 상태
        last_login_at:
          type: string
          format: date-time
          description: 마지막 로그인 일시
    UsersUpdateRequest:
      type: object
      properties:
        login_id:
          type: string
          description: 로그인 ID(사용자 식별자)
        email:
          type: string
          description: 사용자 이메일
        phone:
          type: string
          description: 사용자 휴대폰 번호
        password_hash:
          type: string
          description: 비밀번호 해시
        status:
          type: string
          enum:
            - ACTIVE
            - PENDING
            - DISABLED
          description: 계정 상태
        last_login_at:
          type: string
          format: date-time
          description: 마지막 로그인 일시
        updated_at:
          type: string
          format: date-time
          description: 수정 일시
  responses:
    BadRequestResponse:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerErrorResponse:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ResourceConflictResponse:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ResourceNotFoundResponse:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'